+++ code/cell.electron/app/src/types.ts
@@ -1,3 +1,2 @@
 export * from '@platform/cell.types/lib/types/types.electron';
 export * from './types.g';
-export * from './types.g2'; // TEMP 🐷 - rename
